Week6/Day2/DailyChallenge/requirements.txt
asgiref==3.7.1
Babel==2.12.1
certifi==2023.5.7
charset-normalizer==3.1.0
customtkinter==5.1.3
darkdetect==0.8.0
distlib==0.3.6
Django==4.2.1
django-phonenumber-field==7.1.0
Faker==18.10.1
filelock==3.12.0
idna==3.4
phonenumbers==8.13.11
Pillow==9.5.0
platformdirs==3.5.1
psycopg2-binary==2.9.6
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
urllib3==2.0.2
virtualenv==20.23.0
